+++ README.md
@@ -11,6 +11,7 @@
 - إمكانية تغيير اللغة بين العربية والإنجليزية مع تسجيل كل تغيير في وحدة التحكم.
 - نظام ألوان متناسق قابل للتخصيص مع دعم الوضعين الفاتح والداكن.
 - قائمة مخصصة للهواتف تظهر العناصر في منتصف الشاشة مع بقاء أزرار تغيير المظهر واللغة ثابتة في الشريط العلوي.
+- هيكل ملفات منظم مع فصل مكونات الواجهة في مجلدات **layout** و**sections** لسهولة الصيانة.
 
 ## التقنيات المستخدمة
