+++ src/pages/Index.tsx
@@ -1,14 +1,14 @@
 import React from 'react';
 import { Helmet } from 'react-helmet';
-import Navigation from '@/components/Navigation';
+import Navigation from '@/components/layout/Navigation';
 import HeroSlider from '@/components/HeroSlider';
-import Services from '@/components/Services';
-import Portfolio from '@/components/Portfolio';
-import About from '@/components/About';
-import CTA from '@/components/CTA';
-import Contact from '@/components/Contact';
-import Footer from '@/components/Footer';
-import BackToTopButton from '@/components/BackToTopButton';
+import Services from '@/components/sections/Services';
+import Portfolio from '@/components/sections/Portfolio';
+import About from '@/components/sections/About';
+import CTA from '@/components/sections/CTA';
+import Contact from '@/components/sections/Contact';
+import Footer from '@/components/layout/Footer';
+import BackToTopButton from '@/components/layout/BackToTopButton';
 import { useLanguage } from '@/contexts/LanguageContext';
 
 const Index: React.FC = () => {
