+++ tsconfig.json
@@ -14,6 +14,6 @@
     "skipLibCheck": true,
     "allowJs": true,
     "noUnusedLocals": false,
-    "strictNullChecks": false
+    "strict": true
   }
 }
